verificação de erro da api, mais informações na main screen
