+++ src/App.jsx
@@ -1,15 +1,12 @@
-import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 import Home from './pages/Home';
 
 function App() {
-
   return (
-    <BrowserRouter>
     <Routes>
-      <Route path='/' element={ <Home /> } />
+      <Route path="/" element={<Home />} />
     </Routes>
-    </BrowserRouter>
-  )
+  );
 }
 
-export default App
+export default App;
